code/requirements.txt
slither-analyzer==0.9.1
solc-select==1.0.4